Cleanup of org.springframework.http.client
Moved all package protected request and response classes in http.client
package into their respective factories.
