+++ spring-web/src/main/java/org/springframework/http/client/HttpComponentsAsyncClientHttpRequestFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2015 the original author or authors.
+ * Copyright 2002-2016 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,21 +18,34 @@ package org.springframework.http.client;
 
 import java.io.IOException;
 import java.net.URI;
+import java.util.concurrent.Future;
 
+import org.apache.http.HttpEntity;
+import org.apache.http.HttpEntityEnclosingRequest;
+import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.Configurable;
 import org.apache.http.client.methods.HttpUriRequest;
 import org.apache.http.client.protocol.HttpClientContext;
+import org.apache.http.concurrent.FutureCallback;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.nio.client.CloseableHttpAsyncClient;
 import org.apache.http.impl.nio.client.HttpAsyncClients;
 import org.apache.http.nio.client.HttpAsyncClient;
+import org.apache.http.nio.entity.NByteArrayEntity;
 import org.apache.http.protocol.HttpContext;
 
 import org.springframework.beans.factory.InitializingBean;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.util.Assert;
+import org.springframework.util.concurrent.FailureCallback;
+import org.springframework.util.concurrent.FutureAdapter;
+import org.springframework.util.concurrent.ListenableFuture;
+import org.springframework.util.concurrent.ListenableFutureCallback;
+import org.springframework.util.concurrent.ListenableFutureCallbackRegistry;
+import org.springframework.util.concurrent.SuccessCallback;
 
 /**
  * Asynchronous extension of the {@link HttpComponentsClientHttpRequestFactory}. Uses
@@ -150,4 +163,129 @@ public class HttpComponentsAsyncClientHttpRequestFactory extends HttpComponentsC
 		}
 	}
 
+	static final class HttpComponentsAsyncClientHttpRequest extends AbstractBufferingAsyncClientHttpRequest {
+
+		final HttpAsyncClient httpClient;
+
+		final HttpUriRequest httpRequest;
+
+		final HttpContext httpContext;
+
+
+		HttpComponentsAsyncClientHttpRequest(HttpAsyncClient httpClient, HttpUriRequest httpRequest, HttpContext httpContext) {
+			this.httpClient = httpClient;
+			this.httpRequest = httpRequest;
+			this.httpContext = httpContext;
+		}
+
+
+		@Override
+		public HttpMethod getMethod() {
+			return HttpMethod.resolve(this.httpRequest.getMethod());
+		}
+
+		@Override
+		public URI getURI() {
+			return this.httpRequest.getURI();
+		}
+
+		@Override
+		protected ListenableFuture<ClientHttpResponse> executeInternal(HttpHeaders headers, byte[] bufferedOutput)
+				throws IOException {
+
+			addHeaders(this.httpRequest, headers);
+
+			if (this.httpRequest instanceof HttpEntityEnclosingRequest) {
+				HttpEntityEnclosingRequest entityEnclosingRequest = (HttpEntityEnclosingRequest) this.httpRequest;
+				HttpEntity requestEntity = new NByteArrayEntity(bufferedOutput);
+				entityEnclosingRequest.setEntity(requestEntity);
+			}
+
+			final HttpResponseFutureCallback callback = new HttpResponseFutureCallback();
+			final Future<HttpResponse> futureResponse =
+					this.httpClient.execute(this.httpRequest, this.httpContext, callback);
+			return new ClientHttpResponseFuture(futureResponse, callback);
+		}
+
+
+		private static class HttpResponseFutureCallback implements
+				FutureCallback<HttpResponse> {
+
+			private final ListenableFutureCallbackRegistry<ClientHttpResponse> callbacks =
+					new ListenableFutureCallbackRegistry<ClientHttpResponse>();
+
+			public void addCallback(ListenableFutureCallback<? super ClientHttpResponse> callback) {
+				this.callbacks.addCallback(callback);
+			}
+
+			public void addSuccessCallback(SuccessCallback<? super ClientHttpResponse> callback) {
+				this.callbacks.addSuccessCallback(callback);
+			}
+
+			public void addFailureCallback(FailureCallback callback) {
+				this.callbacks.addFailureCallback(callback);
+			}
+
+			@Override
+			public void completed(HttpResponse result) {
+				this.callbacks.success(new HttpComponentsAsyncClientHttpResponse(result));
+			}
+
+			@Override
+			public void failed(Exception ex) {
+				this.callbacks.failure(ex);
+			}
+
+			@Override
+			public void cancelled() {
+			}
+		}
+
+
+		private static class ClientHttpResponseFuture extends FutureAdapter<ClientHttpResponse, HttpResponse>
+				implements ListenableFuture<ClientHttpResponse> {
+
+			private final HttpResponseFutureCallback callback;
+
+			public ClientHttpResponseFuture(Future<HttpResponse> futureResponse, HttpResponseFutureCallback callback) {
+				super(futureResponse);
+				this.callback = callback;
+			}
+
+			@Override
+			protected ClientHttpResponse adapt(HttpResponse response) {
+				return new HttpComponentsAsyncClientHttpResponse(response);
+			}
+
+			@Override
+			public void addCallback(ListenableFutureCallback<? super ClientHttpResponse> callback) {
+				this.callback.addCallback(callback);
+			}
+
+			@Override
+			public void addCallback(SuccessCallback<? super ClientHttpResponse> successCallback, FailureCallback failureCallback) {
+				this.callback.addSuccessCallback(successCallback);
+				this.callback.addFailureCallback(failureCallback);
+			}
+		}
+
+
+	}
+
+	private static final class HttpComponentsAsyncClientHttpResponse extends
+			HttpComponentsClientHttpResponse {
+
+
+		HttpComponentsAsyncClientHttpResponse(HttpResponse httpResponse) {
+			super(httpResponse);
+		}
+
+
+		@Override
+		public void close() {
+	        // HTTP responses returned by async HTTP client are not bound to an
+	        // active connection and do not have to deallocate any resources...
+		}
+
+	}
 }
