--- spring-web/src/main/java/org/springframework/http/client/HttpComponentsAsyncClientHttpResponse.java
@@ -1,85 +1,0 @@
-/*
- * Copyright 2002-2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.http.client;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-import org.apache.http.Header;
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpResponse;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.util.StreamUtils;
-
-/**
- * {@link ClientHttpResponse} implementation that uses
- * Apache HttpComponents HttpClient to execute requests.
- *
- * <p>Created via the {@link HttpComponentsAsyncClientHttpRequest}.
- *
- * @author Oleg Kalnichevski
- * @author Arjen Poutsma
- * @since 4.0
- * @see HttpComponentsAsyncClientHttpRequest#executeAsync()
- */
-final class HttpComponentsAsyncClientHttpResponse extends AbstractClientHttpResponse {
-
-	private final HttpResponse httpResponse;
-
-	private HttpHeaders headers;
-
-
-	HttpComponentsAsyncClientHttpResponse(HttpResponse httpResponse) {
-		this.httpResponse = httpResponse;
-	}
-
-
-	@Override
-	public int getRawStatusCode() throws IOException {
-		return this.httpResponse.getStatusLine().getStatusCode();
-	}
-
-	@Override
-	public String getStatusText() throws IOException {
-		return this.httpResponse.getStatusLine().getReasonPhrase();
-	}
-
-	@Override
-	public HttpHeaders getHeaders() {
-		if (this.headers == null) {
-			this.headers = new HttpHeaders();
-			for (Header header : this.httpResponse.getAllHeaders()) {
-				this.headers.add(header.getName(), header.getValue());
-			}
-		}
-		return this.headers;
-	}
-
-	@Override
-	public InputStream getBody() throws IOException {
-		HttpEntity entity = this.httpResponse.getEntity();
-		return (entity != null ? entity.getContent() : StreamUtils.emptyInput());
-	}
-
-	@Override
-	public void close() {
-        // HTTP responses returned by async HTTP client are not bound to an
-        // active connection and do not have to deallocate any resources...
-	}
-
-}
