--- spring-web/src/main/java/org/springframework/http/client/SimpleBufferingClientHttpRequest.java
@@ -1,110 +1,0 @@
-/*
- * Copyright 2002-2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.http.client;
-
-import java.io.IOException;
-import java.net.HttpURLConnection;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Map;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.util.FileCopyUtils;
-import org.springframework.util.StringUtils;
-
-/**
- * {@link ClientHttpRequest} implementation that uses standard JDK facilities to
- * execute buffered requests. Created via the {@link SimpleClientHttpRequestFactory}.
- *
- * @author Arjen Poutsma
- * @author Juergen Hoeller
- * @since 3.0
- * @see SimpleClientHttpRequestFactory#createRequest(java.net.URI, HttpMethod)
- */
-final class SimpleBufferingClientHttpRequest extends AbstractBufferingClientHttpRequest {
-
-	private final HttpURLConnection connection;
-
-	private final boolean outputStreaming;
-
-
-	SimpleBufferingClientHttpRequest(HttpURLConnection connection, boolean outputStreaming) {
-		this.connection = connection;
-		this.outputStreaming = outputStreaming;
-	}
-
-
-	@Override
-	public HttpMethod getMethod() {
-		return HttpMethod.resolve(this.connection.getRequestMethod());
-	}
-
-	@Override
-	public URI getURI() {
-		try {
-			return this.connection.getURL().toURI();
-		}
-		catch (URISyntaxException ex) {
-			throw new IllegalStateException("Could not get HttpURLConnection URI: " + ex.getMessage(), ex);
-		}
-	}
-
-	@Override
-	protected ClientHttpResponse executeInternal(HttpHeaders headers, byte[] bufferedOutput) throws IOException {
-		addHeaders(this.connection, headers);
-
-		// JDK <1.8 doesn't support getOutputStream with HTTP DELETE
-		if (HttpMethod.DELETE == getMethod() && bufferedOutput.length == 0) {
-			this.connection.setDoOutput(false);
-		}
-
-		if (this.connection.getDoOutput() && this.outputStreaming) {
-			this.connection.setFixedLengthStreamingMode(bufferedOutput.length);
-		}
-		this.connection.connect();
-		if (this.connection.getDoOutput()) {
-			FileCopyUtils.copy(bufferedOutput, this.connection.getOutputStream());
-		}
-
-		return new SimpleClientHttpResponse(this.connection);
-	}
-
-
-	/**
-	 * Add the given headers to the given HTTP connection.
-	 * @param connection the connection to add the headers to
-	 * @param headers the headers to add
-	 */
-	static void addHeaders(HttpURLConnection connection, HttpHeaders headers) {
-		for (Map.Entry<String, List<String>> entry : headers.entrySet()) {
-			String headerName = entry.getKey();
-			if (HttpHeaders.COOKIE.equalsIgnoreCase(headerName)) {  // RFC 6265
-				String headerValue = StringUtils.collectionToDelimitedString(entry.getValue(), "; ");
-				connection.setRequestProperty(headerName, headerValue);
-			}
-			else {
-				for (String headerValue : entry.getValue()) {
-					String actualHeaderValue = headerValue != null ? headerValue : "";
-					connection.addRequestProperty(headerName, actualHeaderValue);
-				}
-			}
-		}
-	}
-
-}
