--- spring-web/src/main/java/org/springframework/http/client/SimpleStreamingClientHttpRequest.java
@@ -1,105 +1,0 @@
-/*
- * Copyright 2002-2015 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.http.client;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
-import java.net.URI;
-import java.net.URISyntaxException;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.util.StreamUtils;
-
-/**
- * {@link ClientHttpRequest} implementation that uses standard JDK facilities to
- * execute streaming requests. Created via the {@link SimpleClientHttpRequestFactory}.
- *
- * @author Arjen Poutsma
- * @since 3.0
- * @see SimpleClientHttpRequestFactory#createRequest(java.net.URI, HttpMethod)
- */
-final class SimpleStreamingClientHttpRequest extends AbstractClientHttpRequest {
-
-	private final HttpURLConnection connection;
-
-	private final int chunkSize;
-
-	private OutputStream body;
-
-	private final boolean outputStreaming;
-
-
-	SimpleStreamingClientHttpRequest(HttpURLConnection connection, int chunkSize, boolean outputStreaming) {
-		this.connection = connection;
-		this.chunkSize = chunkSize;
-		this.outputStreaming = outputStreaming;
-	}
-
-
-	public HttpMethod getMethod() {
-		return HttpMethod.resolve(this.connection.getRequestMethod());
-	}
-
-	@Override
-	public URI getURI() {
-		try {
-			return this.connection.getURL().toURI();
-		}
-		catch (URISyntaxException ex) {
-			throw new IllegalStateException("Could not get HttpURLConnection URI: " + ex.getMessage(), ex);
-		}
-	}
-
-	@Override
-	protected OutputStream getBodyInternal(HttpHeaders headers) throws IOException {
-		if (this.body == null) {
-			if (this.outputStreaming) {
-				int contentLength = (int) headers.getContentLength();
-				if (contentLength >= 0) {
-					this.connection.setFixedLengthStreamingMode(contentLength);
-				}
-				else {
-					this.connection.setChunkedStreamingMode(this.chunkSize);
-				}
-			}
-			SimpleBufferingClientHttpRequest.addHeaders(this.connection, headers);
-			this.connection.connect();
-			this.body = this.connection.getOutputStream();
-		}
-		return StreamUtils.nonClosing(this.body);
-	}
-
-	@Override
-	protected ClientHttpResponse executeInternal(HttpHeaders headers) throws IOException {
-		try {
-			if (this.body != null) {
-				this.body.close();
-			}
-			else {
-				SimpleBufferingClientHttpRequest.addHeaders(this.connection, headers);
-				this.connection.connect();
-			}
-		}
-		catch (IOException ex) {
-			// ignore
-		}
-		return new SimpleClientHttpResponse(this.connection);
-	}
-
-}
